Update IBS_Estimation_MCMC.R
R script that contains the function to run the IBS, calculate the simulated value of the log-likelihood function, and run the Metropolis-Hastings algorithm to generate the joint posterior distributions of unknown parameters. The starting values, covariance matrix, and matrix scalar, were generated from successive pilot runs described in the Supplemental text.